Board Class를 ChessGame Class와 Board Class로 분리, 각 Class에 맞는 method 리팩토링
