+++ src/test/java/chess/BoardTest.java
@@ -18,20 +18,6 @@ class BoardTest {
     }
 
     @Test
-    public void create() throws Exception {
-        board.initialize();
-        String blankRank = appendNewLine("........");
-        assertEquals(
-                appendNewLine("RNBQKBNR") +
-                        appendNewLine("PPPPPPPP") +
-                        blankRank + blankRank + blankRank + blankRank +
-                        appendNewLine("pppppppp") +
-                        appendNewLine("rnbqkbnr"),
-                board.print()
-        );
-    }
-
-    @Test
     public void verifyPieceCount() {
         board.initialize();
         assertEquals(board.getPieceCount(Piece.createBlackPawn()), 8);
@@ -69,50 +55,6 @@ class BoardTest {
         board.move(position, piece);
 
         assertEquals(piece, board.getPieceAt(position));
-        System.out.println(board.print());
-    }
-
-    @Test
-    public void caculcatePoint() throws Exception {
-        board.initializeEmpty();
-
-        addPiece("b6", Piece.createBlackPawn());
-        addPiece("e6", Piece.createBlackQueen());
-        addPiece("b8", Piece.createBlackKing());
-        addPiece("c8", Piece.createBlackRook());
-
-        addPiece("f2", Piece.createWhitePawn());
-        addPiece("g2", Piece.createWhitePawn());
-        addPiece("e1", Piece.createWhiteRook());
-        addPiece("f1", Piece.createWhiteKing());
-
-        assertEquals(15.0, board.calculateScore(Piece.Color.BLACK), 0.01);
-        assertEquals(7.0, board.calculateScore(Piece.Color.WHITE), 0.01);
-
-        System.out.println(board.print());
-    }
-
-    private void addPiece(String position, Piece piece) {
-        board.move(position, piece);
-    }
-
-    @Test
-    public void sortPieceAscend() {
-//        board.initialize();
-        board.initializeEmpty();
-
-        addPiece("b6", Piece.createBlackPawn());
-        addPiece("e6", Piece.createBlackQueen());
-        addPiece("b8", Piece.createBlackKing());
-        addPiece("c8", Piece.createBlackRook());
-
-        addPiece("f2", Piece.createWhitePawn());
-        addPiece("g2", Piece.createWhitePawn());
-        addPiece("e1", Piece.createWhiteRook());
-        addPiece("f1", Piece.createWhiteKing());
-
-        List<Piece> res = board.sortPiecesByScore(Piece.Color.BLACK, true);
-        for(Piece piece:res) System.out.println(piece.getScore());
     }
 
     @Test
